+++ game/towers/grenadier.py
@@ -73,7 +73,7 @@ class GrenadierProjectile(Entity):
     def __init__(self, location: Location = Location(),
                  priority: int = 0,
                  *,
-                 velocity: tuple[int, int] = (0, 0),
+                 velocity: tuple[float, float] = (0, 0),
                  damage: int = 0,
                  aoe_radius: int = 0):
         super().__init__(location, priority)
@@ -85,11 +85,11 @@ class GrenadierProjectile(Entity):
         self._aoe_radius = aoe_radius
 
     @property
-    def velocity(self) -> tuple[int, int]:
+    def velocity(self) -> tuple[float, float]:
         return self._velocity
 
     @velocity.setter
-    def velocity(self, value: tuple[int, int]):
+    def velocity(self, value: tuple[float, float]):
         self._velocity = value
 
     def tick(self, tick_count: int) -> None:
